Account: extract `LogOut` flow helper
