+++ components/brave_account/BUILD.gn
@@ -19,6 +19,8 @@ source_set("brave_account") {
     "flows/change_password.h",
     "flows/get_service_token.cc",
     "flows/get_service_token.h",
+    "flows/log_out.cc",
+    "flows/log_out.h",
     "flows/login.cc",
     "flows/login.h",
     "flows/register.cc",
@@ -130,6 +132,7 @@ source_set("unit_tests") {
     "brave_account_service_unittest.cc",
     "flows/change_password_unittest.cc",
     "flows/get_service_token_unittest.cc",
+    "flows/log_out_unittest.cc",
     "flows/login_unittest.cc",
     "flows/register_unittest.cc",
     "flows/resend_verification_email_unittest.cc",
